+++ components/slack/src/actions/messages/searchMessages.ts
@@ -10,8 +10,9 @@ export const searchMessages = action({
   },
   perform: async (
     { debug: { enabled: debug } },
-    { connection, count, highlight, page, query, sort, sort_dir, team_id },
+    { connection, pagination, highlight, query, sort, sort_dir, team_id },
   ) => {
+    const { count, page } = pagination;
     debugLogger({
       debug,
       count,